core/pyproject.toml
[project]
name = "awa"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.49.0",
    "fastapi>=0.136.0",
    "greenlet>=3.2.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.14.0",
    "uvicorn>=0.45.0",
    "langgraph>=1.1.9",
    "google-adk>=2.0.0b1",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.21.0",
    "isort>=8.0.1",
    "pre-commit>=4.6.0",
    "pycln>=2.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
  # Transitive authlib import (via google-adk); upstream migration to joserfc.
  "ignore::authlib.deprecate.AuthlibDeprecationWarning",
  # ADK experimental feature flags; not actionable in this repo.
  "ignore::UserWarning:google.adk.features._feature_decorator",
]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # Pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
  "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
# google.adk imports must follow runtime warning filters; FastAPI Depends(Provide[...]) is intentional.
"src/main.py" = ["E402", "B008"]
# DATABASE_URL must be set before importing src.main (see module docstring).
"tests/conftest.py" = ["E402"]

[tool.ruff.format]
quote-style = "double"

[tool.isort]
profile = "black"
line_length = 88
src_paths = ["src", "tests"]
known_first_party = ["src"]

[tool.pycln]
all = true


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv.sources]
src = { path = "src" }
